move rendering_tester to support/
